Fix greyed out applications list on HMI launch
There was a problem that when HMI does not receive
from SDL UpdateDeviceList, it does not set info
about drivers device so all devices, that will be
connected after that will be treated as non-driver
devices and will be displayed as inactive.

To solve that problem was added driver info init
on receiving onAppRegistered notification because
it also contains information about application device.