Fix Uncaught TypeError when receiving SendLocation
There was a problem that NavigationController could
be not inited before receiving SendLocation and this
lead to reported error.

There was added initialize of NavigationController on
requesting Navigation.IsReady so if navigation module
is active, then it will be inited on Navi.isReady request
and result of initialization will be transfered in isReady
response.