Fix SubscribeVehicleData/UnsubscribeVehicleData response
There was a problem that HMI keeps subscription status of
applications on SubscribeVehicleData/UnsubscribeVehicleData
requests is comming. However this is an SDL responsibility
to manage with subscriptions.

In this commit was updated HMI logic to provide only list of
available vehicle data and just check requested data using it.
All logic related to subscriptions were removed.