Created the single websocket instance
There was a problem with multiple an instance
RPCClient object which also creates a multiple instance
of websocket for the different components. When the HMI
receive messages with different interfaces from the SDL
an order of the execution this messages isn't correct.
Message which was send by SDL first can be execute by the
HMI later then second message which was send from the SDL
by the multiple instance of the websocket reason.
The solution is to create a single instance of the
RPCClient object and the other components will contain
just a reference to the RPCClient.