Enable vehicle emulation by default (#239)
After including of feature "[SDL 0221] Remote Control -
Allow Multiple Modules per Module Type" into the release
6.0, new SDL is now working with the SystemCapabilities
with "moduleId" mandatory field. However, "no emulation"
mode on HMI was designed to test older SDL versions which
are not supporting this feature and working without "moduleId".

In order to make latest HMI to be on the same page with SDL,
emulation mode has been enabled by default. This simplifies
work with HMI as tester does not need to change emulation
mode each time on each HMI start. It makes more sense to
disable emulation when it is not needed, but by default it
should be enabled for the latest releases.