Fix URLs in OnSystemRequest after retry
There was noticed an issue that HMI does not include url
parameter into OnSystemRequest notification in EXT mode
As was found, retry sequence was using old structure format
for extracting URL and appid according to GetURLS request
which was in previous releases. However, after introducing
GetPolicyConfigurationData this structure was changed and
not updated accordingly. The following changes were done
to align it with the actual strucure.
Also was removed duplicated js file which is not used in
project and just brings confusions during development.