Feature/sdl 0190 resumption data error handling (#394)
* Add implementation `resumption data error handling` on the HMI

* Change Global RPC processing

For global RPC, the ability to respond with different resultCodes to
different RPC was added. The answers are added to the queue and then
subtracted for each subsequent requests.

* fixed HMI sends UI.SetGlobalProperties_TTS.SetGlobalProperties_responses_with_empty_info_message

* Fix RPC settings icon style

* Fix result code customization for SubscribeVD

Was added result code override if user specified custom code
and do not subscribe to VD if code is negative

* Add CreateWindow result code customization

* Add DO_NOT_RESPOND support

* Fix undefined result code for subscriptions

The issue was that combobox are being filled with result codes
in the later initialization. As a result, value bindings may set
variable to undefined if result codes list is empty. To resolve
that issue, initial element of SUCCESS has been added to the
list before it is updated.

* Add more successful result codes

* fix HMI sends SetGlobalProperties response with info null

* Add check if app contains app_info structure

* Update year in copyright

* Removed unrelated images

* Add separate code customization for choice sets

* Remove SoftButtons limitation for main windows

* Fix capabilities check result codes

Also all result code calculation functions for all interfaces
have been updated to properly consider result codes if
some error was raised after capabilities check

* Fix AddSubMenu after conflicts resolution

Co-authored-by: v-malko4 <vmalkov@luxoft.com>
Co-authored-by: BSolonenko <BSolonenko@luxoft.com>
Co-authored-by: yurii <ypostolov>