Fix nested commands removal (#415)
In case when commands organized in a tree hierarchy there
might be a case when applicaiton deletes command with
childs. In that case SDL removes parent command first and then
all its childs starting from the bottom one. In case when SDL
tries to remove the last child, HMI responds with INVALID_ID
because it can't find that node by menu_id as its parent was
deleted first.
To fix that issue, logic of commands removal was updated so
now HMI responds with SUCCESS if commandsList contain
the array by menuID. It makes more sense because parent
node could be deleted first.