Added return statement for early resolve/reject (#425)
There was noticed that in several cases promise statement
continues to execute even after resolve()/reject() was called
in early if statement. This is because JS has the concept
"run to completion". Unless an error is thrown, a function is
executed until a return statement or its end is reached.
To fix that, return statements have been placed for cases
where resolve/reject is called early. This prevents functions
to execute unnecessary code.
Since the return value of the callback is ignored, we can save a
line by returning the reject/resolve statement.