Fix ScrollableMessage, SetGlobalProperties, AlertManeuver
The was noticed a wrong image validation logic in mentioned
rpcs. In all cases HMI responds twice - first time with
regular result code and the second time with the result of
images validation.
To fix that, image validation has been placed into the proper
place in the code and appropriate breaks were added.
Also, was added timer stop for AlertManeuver popup.