Fix unnecessary button events from bottom view
There was noticed an issue that HMI captures ButtonUp and
click events in case when button on the upper view was clicked,
view closed and there is another button on the lower view. In
that case, unexpected events for these buttons on the lower
view will be raised.
To fix that, buttonDown actions on the VR popup have been
replaced with buttonUp action which prevents view closure
before all events are captured.