Fix UNSUPPORTED_RESOURCE result code for images (#450)
* Fix UNSUPPORTED_RESOURCE result code for images

Make HMI to respond with the proper result code in case when
STATIC image is received. Also removed redundant and
commented code to make things cleaner.

* Fix PerformInteraction image validation

* Fix empty info message and params forwarding

* Fix response params forwarding

There was noticed a regression that HMI does not forward
params in case of error. This is because of complicated and
wrong logic of calculation of some corner cases.
To avoid that, condition was reformatted as a more readable
separate function.

* Fix TTS.Speak empty chunks handling