Fix HMI connection to Policy Manager during PTU
There was noticed a timing issue when HMI is not fast enough
to connect to Policy Manager backend after ignition off/on. As
SDL starts before the Policy Manager, there might be an issue
when SDL continues PTU however HMI is not connected to
PM yet which causes an exception and breaks PTU sequence.
Also, HMI is trying to perform PTU even if no any apps
registered yet which makes no sense if PTU over internal
modem is disabled.
The following changes were done:
- Postpone PTU sequence until at least one app is registered
- Resume PTU sequence once app is registered and PTU was
postponed
- Postpone sending of messages to Policy Manager until
socket is connected
- Resume sending of postponed messages to PM once
coonection is established